Refactor/vse optimization

+++ botfront/imports/api/story/stories.methods.js
@@ -22,17 +22,18 @@ Meteor.methods({
         return Stories.insert({ ...story, events: aggregateEvents(story) });
     },
 
-    async 'stories.update'(story, projectId) {
+    async 'stories.update'(story, projectId, options = {}) {
         check(story, Object);
         check(projectId, String);
+        check(options, Object);
+        const { noClean } = options;
         const {
             _id, path, ...rest
         } = story;
-        if (!path) {
-            return Stories.update({ _id }, { $set: { ...rest } });
-        }
+        
+        if (!path) return Stories.update({ _id }, { $set: { ...rest } });
         const originStory = Stories.findOne({ _id });
-        const { events: oldEvents } = originStory || {};
+
         // passing story.story and path[(last index)] AKA storyBranchId to aggregate events allows it to aggregate events with the updated story md
         const newEvents = aggregateEvents(originStory, { ...rest, _id: path[path.length - 1] }); // path[(last index)] is the id of the updated branch
 
@@ -47,9 +48,11 @@ Meteor.methods({
             : rest;
         const result = await Stories.update({ _id }, { $set: { ...update, events: newEvents } });
 
-        // check if a response was removed
-        const removedEvents = (oldEvents || []).filter(event => event.match(/^utter_/) && !newEvents.includes(event));
-        deleteResponsesRemovedFromStories(removedEvents, projectId);
+        if (!noClean) { // check if a response was removed
+            const { events: oldEvents } = originStory || {};
+            const removedEvents = (oldEvents || []).filter(event => event.match(/^utter_/) && !newEvents.includes(event));
+            deleteResponsesRemovedFromStories(removedEvents, projectId);
+        }
         return result;
     },
